pangrams: rename class to PangramChecker
