+++ exercises/pangram/src/main/java/PangramChecker.java
@@ -1,4 +1,4 @@
-public class Pangrams {
+public class PangramChecker {
 
     public boolean isPangram(String input) {
         throw new UnsupportedOperationException("Method has not been implemented yet.");
